+++ src/main/java/kitchenpos/product/applicatiion/ProductService.java
@@ -1,7 +1,9 @@
-package kitchenpos.application;
+package kitchenpos.product.applicatiion;
 
 import kitchenpos.domain.*;
 import kitchenpos.infra.PurgomalumClient;
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.domain.ProductRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
