+++ src/main/java/kitchenpos/product/domain/ProductRepository.java
@@ -1,8 +1,9 @@
-package kitchenpos.domain;
+package kitchenpos.product.domain;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import kitchenpos.product.domain.Product;
 
 public interface ProductRepository {
     Product save(Product product);
